goccina.com/requirements.txt
dj-static==0.0.6
Django==1.9
django-modeltranslation==0.11
Pillow==2.8.1
python-slugify==1.2.1
static3==0.7.0
Unidecode==0.4.19
